+++ tests/integration/async/info_test.py
@@ -47,7 +47,7 @@ async def secondary_index(session):
     try:
         task = await (
             session.index(namespace, set_name)
-            .on_bin("idx_bin").named(index_name).numeric().create()
+            .on_bin("idx_bin").named(index_name).integer().create()
         )
     except Exception:
         # Left behind by an interrupted run: its build is done, no task to await.
